Establish initial index.html and styling
Created a standardized page layout using a top-level nav, a page
header/hero section, a main section for page contents, and a footer.
Each of these sections is utilizing varying degrees of flexbox to
both properly place elements on the page as well as support a
reasonably flexible site design.