Add landing page content and partly style
Began adding website content to the landing page. As content was added,
styling was also added. Further cleanup can be done, like creating
generic classes for re-used styles. This would be better off done now,
at the early stages of the project than later when implementing such a
change would be more tedious.

Items still needed are a logo and testimonial video to be included on
the landing page, as well as any social network links/third party links
that would need to be featured.